pyproject.toml
[tool.pdm]
[tool.pdm.dev-dependencies]
dev = [
    "flake8>=6.0.0",
    "autopep8>=2.0.2",
]

[project]
name = "yutservice"
version = "0.0.1"
description = ""
authors = [
    {name = "ReRubis", email = "slesarenkod@gmail.com"},
]
dependencies = [
    "asyncio>=3.4.3",
    "aiohttp>=3.8.4",
    "aiofiles>=23.1.0",
    "fastapi[all]>=0.95.0",
    "sqlalchemy>=2.0.7",
    "google-api-python-client>=2.81.0",
    "sqlalchemy-utils>=0.40.0",
    "psycopg2>=2.9.5",
    "yt-dlp>=2023.10.13",
    "python-jose>=3.3.0",
    "bcrypt>=4.0.1",
]
requires-python = ">=3.11"
license = {text = "MIT"}
